feat: modern-js switch on rspack
