--- packages/solutions/app-tools/src/builder/builder-webpack/loaders/serverModuleLoader.ts
@@ -1,7 +1,0 @@
-import { LoaderContext } from 'webpack';
-
-function loader(this: LoaderContext<void>) {
-  return `module.exports = {}`;
-}
-
-export default loader;
