+++ packages/solutions/app-tools/src/builder/builder-webpack/webpackPlugins/RouterPlugin.ts
