+++ packages/solutions/app-tools/src/builder/builder-webpack/webpackPlugins/index.ts
@@ -1,0 +1,1 @@
+export * from './RouterPlugin';
