+++ packages/solutions/app-tools/src/builder/generator/getBuilderTargets.ts
@@ -1,6 +1,6 @@
 import type { BuilderTarget } from '@modern-js/builder-shared';
 import { isProd, isSSR, isUseSSRBundle } from '@modern-js/utils';
-import type { AppNormalizedConfig } from '@/types';
+import type { AppNormalizedConfig } from '../../types';
 
 export function getBuilderTargets(normalizedConfig: AppNormalizedConfig) {
   const targets: BuilderTarget[] = ['web'];
