+++ packages/solutions/app-tools/src/builder/shared/loaders/serverModuleLoader.ts
@@ -1,0 +1,8 @@
+import type { webpack } from '@modern-js/builder-webpack-provider';
+import type { rspack } from '@modern-js/builder-rspack-provider';
+
+function loader(this: webpack.LoaderContext<void> | rspack.LoaderContext) {
+  return `module.exports = {}`;
+}
+
+export default loader;
