+++ packages/solutions/app-tools/src/types/hooks.ts
@@ -14,6 +14,7 @@ import type {
 } from '@modern-js/types';
 import type { RegisterBuildPlatformResult, DevToolData } from '@modern-js/core';
 import type { Stats, MultiStats } from '@modern-js/builder-shared';
+import type { rspack } from '@modern-js/builder-rspack-provider';
 
 export interface ImportSpecifier {
   local?: string;
@@ -76,16 +77,16 @@ export type AppToolsHooks = {
   beforeDev: AsyncWorkflow<void, unknown>;
   afterDev: AsyncWorkflow<void, unknown>;
   beforeCreateCompiler: AsyncWorkflow<
-    { bundlerConfigs: webpack.Configuration[] },
+    { bundlerConfigs: webpack.Configuration[] | rspack.Configuration[] },
     unknown
   >;
   afterCreateCompiler: AsyncWorkflow<
-    { compiler?: webpack.Compiler | webpack.MultiCompiler },
+    { compiler?: webpack.Compiler | webpack.MultiCompiler | rspack.Compiler },
     unknown
   >;
   beforePrintInstructions: AsyncWaterfall<{ instructions: string }>;
   beforeBuild: AsyncWorkflow<
-    { bundlerConfigs?: webpack.Configuration[] },
+    { bundlerConfigs?: webpack.Configuration[] | rspack.Configuration[] },
     unknown
   >;
   afterBuild: AsyncWorkflow<{ stats?: Stats | MultiStats }, unknown>;
